requirements.txt
pytest
pytest-asyncio
psutil
aiofiles
aiohttp
async-timeout
grappa
PyYAML
uvloop
git+https://github.com/webrecorder/simplechrome.git
attrs
black
sauceclient
selenium
Jinja2